+++ janitor/utils.py
@@ -28,6 +28,7 @@ from pandas.api.types import (
     is_extension_array_dtype,
     is_list_like,
     is_scalar,
+    is_string_dtype,
 )
 from pandas.core.common import apply_if_callable
 
@@ -1593,12 +1594,10 @@ def _data_checks_pivot_wider(
     names_from,
     values_from,
     names_sort,
+    levels_order,
     flatten_levels,
-    names_from_position,
-    names_prefix,
     names_sep,
-    aggfunc,
-    fill_value,
+    names_glue,
 ):
 
     """
@@ -1611,7 +1610,7 @@ def _data_checks_pivot_wider(
 
     if index is not None:
         if is_list_like(index):
-            index = list(index)
+            index = [*index]
         index = _select_columns(index, df)
 
     if names_from is None:
@@ -1620,38 +1619,26 @@ def _data_checks_pivot_wider(
         )
 
     if is_list_like(names_from):
-        names_from = list(names_from)
+        names_from = [*names_from]
     names_from = _select_columns(names_from, df)
 
     if values_from is not None:
-        check("values_from", values_from, [list, str])
+        if is_list_like(values_from):
+            values_from = [*values_from]
         values_from = _select_columns(values_from, df)
 
     check("names_sort", names_sort, [bool])
 
+    if levels_order is not None:
+        check("levesl_order", levels_order, [list])
+
     check("flatten_levels", flatten_levels, [bool])
-
-    if names_from_position is not None:
-        check("names_from_position", names_from_position, [str])
-        if names_from_position not in ("first", "last"):
-            raise ValueError(
-                """
-                The position of `names_from`
-                must be either "first" or "last".
-                """
-            )
-
-    if names_prefix is not None:
-        check("names_prefix", names_prefix, [str])
 
     if names_sep is not None:
         check("names_sep", names_sep, [str])
 
-    if aggfunc is not None:
-        check("aggfunc", aggfunc, [str, list, dict, callable])
-
-    if fill_value is not None:
-        check("fill_value", fill_value, [int, float, str])
+    if names_glue is not None:
+        check("names_glue", names_glue, [callable])
 
     return (
         df,
@@ -1659,12 +1646,10 @@ def _data_checks_pivot_wider(
         names_from,
         values_from,
         names_sort,
+        levels_order,
         flatten_levels,
-        names_from_position,
-        names_prefix,
         names_sep,
-        aggfunc,
-        fill_value,
+        names_glue,
     )
 
 
@@ -1674,12 +1659,10 @@ def _computations_pivot_wider(
     names_from: Optional[Union[List, str]] = None,
     values_from: Optional[Union[List, str]] = None,
     names_sort: Optional[bool] = False,
+    levels_order:Optional[list]=None,
     flatten_levels: Optional[bool] = True,
-    names_from_position: Optional[str] = "first",
-    names_prefix: Optional[str] = None,
-    names_sep: Optional[str] = "_",
-    aggfunc: Optional[Union[str, list, dict, Callable]] = None,
-    fill_value: Optional[Union[int, float, str]] = None,
+    names_sep="_",
+    names_glue: Callable = None,
 ) -> pd.DataFrame:
     """
     This is the main workhorse of the `pivot_wider` function.
@@ -1692,85 +1675,102 @@ def _computations_pivot_wider(
     A dataframe is returned.
     """
 
-    if not names_sort:
+    names_from_all_strings = df.filter(names_from).agg(is_string_dtype).all()
+
+    if names_sort is True:
         # Categorical dtypes created only for `names_from`
         # since that is what will become the new column names
         dtypes = {
             column_name: CategoricalDtype(
-                categories=column.dropna().unique(), ordered=True
+                df[column_name].factorize(sort=False)[-1], ordered=True
             )
-            if column.hasnans
-            else CategoricalDtype(categories=column.unique(), ordered=True)
-            for column_name, column in df.filter(names_from).items()
+            for column_name in names_from
         }
-
         df = df.astype(dtypes)
 
-    if aggfunc is None:
-        df = df.pivot(  # noqa: PD010
-            index=index, columns=names_from, values=values_from
+    new_index = None
+    if values_from is None:
+        if index is None:
+            # existing index label is used as index
+            values_from = df.columns.difference(names_from)
+            if values_from.empty:
+                warnings.warn(
+                    f"""
+                    There are no column labels associated with
+                    `values_from`. An empty dataframe might be
+                    returned.
+                    """,
+                    UserWarning,
+                    stacklevel=2,
+                )
+            values_from = [*values_from]
+            df = df.set_index(names_from, append=True)
+        else:
+            new_index = index + names_from
+            values_from = df.columns.difference(new_index)
+            if values_from.empty:
+                warnings.warn(
+                    f"""
+                    There are no column labels associated with
+                    `values_from`. An empty dataframe might be
+                    returned.
+                    """,
+                    UserWarning,
+                    stacklevel=2,
+                )
+            values_from = [*values_from]
+            df = df.set_index(new_index, append=False)
+    else:
+        if index is None:
+            # existing index label is used as index
+            df = df.set_index(names_from, append=True)
+        else:
+            new_index = index + names_from
+            df = df.set_index(new_index, append=False)
+        df = df.filter(values_from)
+
+    # check if the index is unique:
+    if not df.index.is_unique:
+        raise ValueError(
+            """
+            The dataframe cannot be pivoted
+            to wide form,
+            because the combination of
+            `index` and `names_from`
+            contain duplicates.
+            """
         )
 
-    else:
-        if index:
-            df = df.set_index(index + names_from)
-        else:
-            df = df.set_index(names_from, append=True)
-
-        if values_from:
-            df = df.groupby(
-                level=[*range(df.index.nlevels)],
-                observed=True,
-                dropna=False,
-                sort=False,
-            )[values_from].agg(aggfunc)
-        else:
-            df = df.groupby(
-                level=[*range(df.index.nlevels)],
-                observed=True,
-                dropna=False,
-                sort=False,
-            ).agg(aggfunc)
-
-        df = df.unstack(level=names_from)  # noqa: PD010
-
-    if fill_value is not None:
-        df = df.fillna(fill_value)
+    df = df.unstack(names_from)
 
     # no point keeping `values_from`
     # if it's just one name;
-    # could do same for aggfunc; but not worth the extra check
-    if df.columns.get_level_values(0).unique().size == 1:
+    if (
+        (isinstance(df.columns, pd.MultiIndex))
+        and (df.columns.nlevels > len(names_from))
+        and (len(values_from) == 1)
+    ):
         df = df.droplevel(0, axis="columns")
+    if levels_order:
+        df = df.reorder_levels(order=levels_order, axis='columns')
+        df = df.sort_index(level=0, axis='columns', sort_remaining=False)
 
-    other_levels = None
-    names_from_levels = None
-    df_columns = df.columns
-    df_columns_names = df_columns.names
-    if names_from_position == "first":
-        if df_columns_names[: len(names_from)] != names_from:
-            other_levels = [
-                num
-                for num, name in enumerate(df_columns_names)
-                if name not in names_from
-            ]
-            names_from_levels = [
-                num
-                for num, name in enumerate(df_columns_names)
-                if name in names_from
-            ]
-            df = df.reorder_levels(
-                names_from_levels + other_levels, axis="columns"
-            )
-
-    if not flatten_levels:
+    # an empty df is likely because
+    # there are no `values_from`
+    if any((df.empty, flatten_levels is False)):
         return df
 
-    if df_columns.nlevels > 1:
-        df.columns = [names_sep.join(column_tuples) for column_tuples in df]
+    # ensure all entries in names_from are strings
+    if not names_from_all_strings:
+        new_columns = [tuple(map(str, ent)) for ent in df]
+        df.columns = pd.MultiIndex.from_tuples(new_columns)
 
-    if names_prefix:
-        df = df.add_prefix(names_prefix)
+    if names_glue:
+        names_sep = None
+        df.columns = df.columns.map(names_glue)  
+
+    if names_sep:
+        df.columns = df.columns.map(names_sep.join)
 
     # if columns are of category type
     # this returns columns to object dtype